requirements.txt
matplotlib>=3.5
PyYAML>=6.0
